+++ common/darwin/Classes/FlutterRTCMediaStream.m
@@ -43,11 +43,13 @@ typedef void (^NavigatorUserMediaErrorCallback)(NSString* errorType, NSString* e
  */
 typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
 
+- (NSDictionary*)defaultVideoConstraints {
+    return @{@"minWidth" : @"1280", @"minHeight" : @"720", @"minFrameRate" : @"30"};
+}
+
 - (RTCMediaConstraints*)defaultMediaStreamConstraints {
-  NSDictionary* mandatoryConstraints =
-      @{@"minWidth" : @"1280", @"minHeight" : @"720", @"minFrameRate" : @"30"};
   RTCMediaConstraints* constraints =
-      [[RTCMediaConstraints alloc] initWithMandatoryConstraints:mandatoryConstraints
+      [[RTCMediaConstraints alloc] initWithMandatoryConstraints:[self defaultVideoConstraints]
                                             optionalConstraints:nil];
   return constraints;
 }
@@ -100,9 +102,11 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
     }
   }
 
+#if !defined(TARGET_OS_IPHONE)
   if (audioDeviceId != nil) {
     [self selectAudioInput:audioDeviceId result:nil];
   }
+#endif
 
   NSString* trackId = [[NSUUID UUID] UUIDString];
   RTCAudioTrack* audioTrack = [self.peerConnectionFactory audioTrackWithTrackId:trackId];
@@ -355,6 +359,10 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
         videoDevice = [self findDeviceForPosition:position];
       }
     }
+  }
+
+  if ([videoConstraints isKindOfClass:[NSNumber class]]) {
+    videoConstraints = @{@"mandatory": [self defaultVideoConstraints]};
   }
 
   NSInteger targetWidth = 0;
@@ -620,22 +628,22 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
     }];
   }
 #if TARGET_OS_IPHONE
+
   RTCAudioSession* session = [RTCAudioSession sharedInstance];
   for (AVAudioSessionPortDescription* port in session.session.availableInputs) {
     // NSLog(@"input portName: %@, type %@", port.portName,port.portType);
     [sources addObject:@{
-      @"facing" : @"",
       @"deviceId" : port.UID,
       @"label" : port.portName,
       @"groupId" : port.portType,
       @"kind" : @"audioinput",
     }];
   }
+
   for (AVAudioSessionPortDescription* port in session.currentRoute.outputs) {
     // NSLog(@"output portName: %@, type %@", port.portName,port.portType);
     if (session.currentRoute.outputs.count == 1 && ![port.UID isEqualToString:@"Speaker"]) {
       [sources addObject:@{
-        @"facing" : @"",
         @"deviceId" : @"Speaker",
         @"label" : @"Speaker",
         @"groupId" : @"Speaker",
@@ -643,7 +651,6 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
       }];
     }
     [sources addObject:@{
-      @"facing" : @"",
       @"deviceId" : port.UID,
       @"label" : port.portName,
       @"groupId" : port.portType,
@@ -657,7 +664,6 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
   NSArray* inputDevices = [audioDeviceModule inputDevices];
   for (RTCIODevice* device in inputDevices) {
     [sources addObject:@{
-      @"facing" : @"",
       @"deviceId" : device.deviceId,
       @"label" : device.name,
       @"kind" : @"audioinput",
@@ -667,7 +673,6 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
   NSArray* outputDevices = [audioDeviceModule outputDevices];
   for (RTCIODevice* device in outputDevices) {
     [sources addObject:@{
-      @"facing" : @"",
       @"deviceId" : device.deviceId,
       @"label" : device.name,
       @"kind" : @"audiooutput",
@@ -900,7 +905,10 @@ typedef void (^NavigatorUserMediaSuccessCallback)(RTCMediaStream* mediaStream);
       return device;
     }
   }
-  return captureDevices[0];
+  if(captureDevices.count > 0) {
+    return captureDevices[0];
+  }
+  return nil;
 }
 
 - (AVCaptureDeviceFormat*)selectFormatForDevice:(AVCaptureDevice*)device