+++ lib/src/native/media_stream_impl.dart
@@ -7,8 +7,7 @@ import 'media_stream_track_impl.dart';
 import 'utils.dart';
 
 class MediaStreamNative extends MediaStream {
-  MediaStreamNative(String streamId, String ownerTag)
-      : super(streamId, ownerTag);
+  MediaStreamNative(super.streamId, super.ownerTag);
 
   factory MediaStreamNative.fromMap(Map<dynamic, dynamic> map) {
     return MediaStreamNative(map['streamId'], map['ownerTag'])