+++ CHANGELOG.md
@@ -1,6 +1,46 @@
 # Changelog
 
 --------------------------------------------
+[0.11.4] - 2024-07-19
+
+* [Android] fix issue for camera switching.
+
+[0.11.3] - 2024-07-12
+
+* Bump version for libwebrtc.
+
+[0.11.2] - 2024-07-09
+
+* [Windows] fix crash for windows.
+* [Darwin] bump WebRTC version for darwin.
+
+[0.11.1] - 2024-06-17
+
+* [macOS] Downgrade macOS system dependencies to 10.14.
+
+[0.11.0] - 2024-06-17
+
+* [Native] upgrade libwebrtc to m125.6422.
+
+[0.10.8] - 2024-06-05
+
+* [iOS] fix(platform_view): fit cover works wrong (#1593)
+* [iOS/macOS] fix: Fix the issue that the video is not displayed when using 'video': true (#1592)
+* [Web] bump dart_webrtc to 1.4.6.
+
+[0.10.7] - 2024-05-30
+
+* [iOS] feat: add PlatformView Renderer for iOS. (#1569)
+* [iOS] fix: audio session control for iOS. (#1590)
+
+[0.10.6] - 2024-05-13
+
+* [Web] Some important fixes for web.
+
+[0.10.5] - 2024-05-13
+
+* [Android] fix: make MediaDeviceInfo (Audio deviceId, label, groupId) consistent. (#1583)
+
 [0.10.4] - 2024-05-06
 
 * [iOS/macOS] chore: update swift webrtc sdks to 114.5735.10 (#1576)